pyproject.toml
[project]
name = "mcp-remote"
version = "0.1.0"
description = "A FastMCP server for searching keywords in Excel files and returning results in JSON format"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "mcp[cli]>=1.9.3",
    "nest-asyncio>=1.6.0",
    "openai>=1.97.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.0",
]
